Travel Itinerary Planning and Management Tool/requirements.txt
ttkbootstrap==1.10.1
Pillow==8.3.2
